# openapi.json
{
  "components": {
    "schemas": {
      "Body_upload_document_api_v1_documents_upload_post": {
        "properties": {
          "doc_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Doc Type"
          },
          "file": {
            "format": "binary",
            "title": "File",
            "type": "string"
          },
          "knowledge_base_id": {
            "default": "default",
            "title": "Knowledge Base Id",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_document_api_v1_documents_upload_post",
        "type": "object"
      },
      "Body_upload_text_api_v1_documents_upload_text_post": {
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "doc_type": {
            "default": "markdown",
            "title": "Doc Type",
            "type": "string"
          },
          "knowledge_base_id": {
            "default": "default",
            "title": "Knowledge Base Id",
            "type": "string"
          },
          "source": {
            "default": "manual_input",
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "content"
        ],
        "title": "Body_upload_text_api_v1_documents_upload_text_post",
        "type": "object"
      },
...
      "ChunkItem": {
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kb_id": {
            "title": "Kb Id",
            "type": "string"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "content",
          "source",
          "type",
          "kb_id"
        ],
        "title": "ChunkItem",
        "type": "object"
      },
      "DocumentStatsResponse": {
        "properties": {
          "collection_name": {
            "title": "Collection Name",
            "type": "string"
          },
          "knowledge_base_count": {
            "title": "Knowledge Base Count",
            "type": "integer"
          },
          "knowledge_base_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Knowledge Base Id"
          },
          "total_documents": {
            "title": "Total Documents",
            "type": "integer"
          }
        },
        "required": [
          "total_documents",
          "collection_name",
          "knowledge_base_count"
        ],
        "title": "DocumentStatsResponse",
        "type": "object"
      },
      "DocumentUploadResponse": {
        "properties": {
          "chunks_added": {
            "title": "Chunks Added",
            "type": "integer"
          },
          "file_id": {
            "description": "Content-addressed id: re-uploading identical bytes yields the same value",
            "title": "File Id",
            "type": "string"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "knowledge_base_id": {
            "title": "Knowledge Base Id",
            "type": "string"
          },
          "knowledge_base_name": {
            "title": "Knowledge Base Name",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "total_docs": {
            "title": "Total Docs",
            "type": "integer"
          }
        },
        "required": [
          "file_id",
          "filename",
          "knowledge_base_id",
          "knowledge_base_name",
          "chunks_added",
          "total_docs",
          "status"
        ],
        "title": "DocumentUploadResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "KnowledgeBaseCreate": {
        "properties": {
          "description": {
            "default": "",
            "maxLength": 200,
            "title": "Description",
            "type": "string"
          },
          "name": {
            "maxLength": 60,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "KnowledgeBaseCreate",
        "type": "object"
      },
      "KnowledgeBaseInfo": {
        "properties": {
          "chunk_count": {
            "default": 0,
            "description": "Indexed passages in this base",
            "title": "Chunk Count",
            "type": "integer"
          },
          "created_at": {
            "default": "",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "default": "",
            "title": "Description",
            "type": "string"
          },
          "document_count": {
            "default": 0,
            "description": "Distinct source files in this base",
            "title": "Document Count",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "is_default": {
            "default": false,
            "title": "Is Default",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ],
        "title": "KnowledgeBaseInfo",
        "type": "object"
      },
      "KnowledgeBaseUpdate": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 60,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "title": "KnowledgeBaseUpdate",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "跨领域科研问答助手：多知识库 RAG 检索、引用溯源与 LangGraph 工作流",
    "title": "AI 科研助手",
    "version": "3.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/documents/chunks": {
      "get": {
        "operationId": "list_chunks_api_v1_documents_chunks_get",
        "parameters": [
          {
            "in": "query",
            "name": "knowledge_base_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Knowledge Base Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 500,
              "maximum": 2000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-Admin-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ChunkItem"
                  },
                  "title": "Response List Chunks Api V1 Documents Chunks Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Chunks",
        "tags": [
          "Documents"
        ]
      }
    },
    "/api/v1/documents/clear": {
      "delete": {
        "description": "Drop every passage in every base. Base names and descriptions survive.\n\nKept for operations and tests; the UI removes one base at a time, which is\nwhat a user actually wants and does not require re-creating the bases they\nstill intend to keep.",
        "operationId": "clear_documents_api_v1_documents_clear_delete",
        "parameters": [
          {
            "in": "header",
            "name": "X-Admin-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Clear Documents",
        "tags": [
          "Documents"
        ]
      }
    },
    "/api/v1/documents/knowledge-bases": {
      "get": {
        "description": "Every base, with its passage and file counts.\n\nCounts are aggregated in one metadata sweep rather than a query per base, so\nthe sidebar renders from a single round trip.",
        "operationId": "list_knowledge_bases_api_v1_documents_knowledge_bases_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/KnowledgeBaseInfo"
                  },
                  "title": "Response List Knowledge Bases Api V1 Documents Knowledge Bases Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Knowledge Bases",
        "tags": [
          "Documents"
        ]
      },
      "post": {
        "operationId": "create_knowledge_base_api_v1_documents_knowledge_bases_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Admin-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeBaseCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeBaseInfo"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Knowledge Base",
        "tags": [
          "Documents"
        ]
      }
    },
    "/api/v1/documents/knowledge-bases/{kb_id}": {
      "delete": {
        "description": "Delete a base and every passage in it.",
        "operationId": "delete_knowledge_base_api_v1_documents_knowledge_bases__kb_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "kb_id",
            "required": true,
            "schema": {
              "title": "Kb Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Admin-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Knowledge Base",
        "tags": [
          "Documents"
        ]
      },
      "patch": {
        "operationId": "update_knowledge_base_api_v1_documents_knowledge_bases__kb_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "kb_id",
            "required": true,
            "schema": {
              "title": "Kb Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Admin-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeBaseUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeBaseInfo"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Knowledge Base",
        "tags": [
          "Documents"
        ]
      }
    },
    "/api/v1/documents/knowledge-bases/{kb_id}/documents": {
      "delete": {
        "description": "Remove every passage in one base, keeping the base itself.\n\nNeeded because the built-in base cannot be deleted — without this, the\nsample corpus that ships with the project could never be cleared out.",
        "operationId": "clear_knowledge_base_api_v1_documents_knowledge_bases__kb_id__documents_delete",
        "parameters": [
          {
            "in": "path",
            "name": "kb_id",
            "required": true,
            "schema": {
              "title": "Kb Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Admin-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Clear Knowledge Base",
        "tags": [
          "Documents"
        ]
      }
    },
    "/api/v1/documents/stats": {
      "get": {
        "operationId": "get_stats_api_v1_documents_stats_get",
        "parameters": [
          {
            "in": "query",
            "name": "knowledge_base_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Knowledge Base Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentStatsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Stats",
        "tags": [
          "Documents"
        ]
      }
    },
    "/api/v1/documents/upload": {
      "post": {
        "operationId": "upload_document_api_v1_documents_upload_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Admin-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_document_api_v1_documents_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentUploadResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Document",
        "tags": [
          "Documents"
        ]
      }
    },
    "/api/v1/documents/upload-text": {
      "post": {
        "operationId": "upload_text_api_v1_documents_upload_text_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Admin-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_text_api_v1_documents_upload_text_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentUploadResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Text",
        "tags": [
          "Documents"
        ]
      }
    },
...
  }
}